+++ spring-cloud-starter-tencent-polaris-discovery/src/main/java/com/tencent/cloud/polaris/extend/consul/ConsulDiscoveryUtil.java
@@ -15,17 +15,16 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package com.tencent.cloud.polaris.tsf.util;
+package com.tencent.cloud.polaris.extend.consul;
 
 import java.util.Map;
 
 import com.ecwid.consul.v1.agent.model.NewService;
-import com.tencent.cloud.common.util.AddressUtils;
 import com.tencent.cloud.common.util.JacksonUtils;
-import com.tencent.cloud.polaris.context.tsf.config.TsfCoreProperties;
-import com.tencent.cloud.polaris.tsf.TsfDiscoveryProperties;
-import com.tencent.cloud.polaris.tsf.TsfHeartbeatProperties;
+import com.tencent.cloud.polaris.context.config.extend.tsf.TsfCoreProperties;
 import com.tencent.polaris.api.config.Configuration;
+import com.tencent.polaris.api.utils.IPAddressUtils;
+import com.tencent.polaris.api.utils.StringUtils;
 import com.tencent.polaris.factory.config.global.ServerConnectorConfigImpl;
 import com.tencent.polaris.plugins.connector.common.constant.ConsulConstant;
 import org.slf4j.Logger;
@@ -39,38 +38,39 @@ import org.springframework.cloud.commons.util.IdUtils;
 import org.springframework.context.ApplicationContext;
 import org.springframework.core.env.Environment;
 import org.springframework.util.Assert;
-import org.springframework.util.StringUtils;
 
 /**
  * TSF registration utils.
  *
  * @author Haotian Zhang
  */
-public final class RegistrationUtil {
+public final class ConsulDiscoveryUtil {
 	/**
 	 * - 分隔符.
 	 */
 	public static final char SEPARATOR = '-';
+
 	/**
 	 * Server connector ID.
 	 */
 	public static final String ID = "consul";
-	private static final Logger LOGGER = LoggerFactory.getLogger(RegistrationUtil.class);
 
-	private RegistrationUtil() {
+	private static final Logger LOGGER = LoggerFactory.getLogger(ConsulDiscoveryUtil.class);
+
+	private ConsulDiscoveryUtil() {
 	}
 
-	public static String getAppName(TsfDiscoveryProperties properties, Environment env) {
+	public static String getAppName(ConsulDiscoveryProperties properties, Environment env) {
 		String appName = properties.getServiceName();
-		if (!StringUtils.isEmpty(appName)) {
+		if (StringUtils.isNotBlank(appName)) {
 			return appName;
 		}
 		return env.getProperty("spring.application.name", "application");
 	}
 
-	public static String getInstanceId(TsfCoreProperties properties, ApplicationContext context) {
+	public static String getInstanceId(ConsulDiscoveryProperties properties, ApplicationContext context) {
 		// tsf consul 不支持 dns，所以这里不需要 normalize，并且由于优雅下线，readiness probe 联动都是依赖 service id 的，normalize 后两边对不上，所以需要去掉 normalize
-		if (!StringUtils.hasText(properties.getInstanceId())) {
+		if (StringUtils.isBlank(properties.getInstanceId())) {
 			return IdUtils.getDefaultInstanceId(context.getEnvironment(), false);
 		}
 		else {
@@ -78,17 +78,7 @@ public final class RegistrationUtil {
 		}
 	}
 
-	// 更新了判断逻辑，仅判断为空的场景
 	public static String normalizeForDns(String s) {
-		// if (s == null || !Character.isLetter(s.charAt(0))
-		// || !Character.isLetterOrDigit(s.charAt(s.length()-1))) {
-		// throw new IllegalArgumentException("Consul service ids must not be empty,
-		// must start with a letter, end with a letter or digit, and have as interior
-		// characters only letters, digits, and hyphen");
-		// }
-
-		// tsf not check consul service instance id start with letter and end with
-		// letter or digit
 		if (s == null) {
 			throw new IllegalArgumentException("Consul service ids must not be empty");
 		}
@@ -113,8 +103,8 @@ public final class RegistrationUtil {
 	}
 
 	public static void setCheck(AutoServiceRegistrationProperties autoServiceRegistrationProperties,
-			TsfDiscoveryProperties properties, TsfCoreProperties tsfCoreProperties, ApplicationContext context,
-			TsfHeartbeatProperties tsfHeartbeatProperties, Registration registration, Configuration configuration) {
+			ConsulDiscoveryProperties properties, TsfCoreProperties tsfCoreProperties, ApplicationContext context,
+			ConsulHeartbeatProperties consulHeartbeatProperties, Registration registration, Configuration configuration) {
 		if (properties.isRegisterHealthCheck()) {
 			Integer checkPort;
 			if (shouldRegisterManagement(autoServiceRegistrationProperties, properties, context)) {
@@ -128,7 +118,7 @@ public final class RegistrationUtil {
 			for (ServerConnectorConfigImpl config : configuration.getGlobal().getServerConnectors()) {
 				if (org.apache.commons.lang.StringUtils.equals(config.getId(), ID)) {
 					Map<String, String> metadata = config.getMetadata();
-					NewService.Check check = createCheck(checkPort, tsfHeartbeatProperties, properties, tsfCoreProperties);
+					NewService.Check check = createCheck(checkPort, consulHeartbeatProperties, properties, tsfCoreProperties);
 					String checkJson = JacksonUtils.serialize2Json(check);
 					LOGGER.debug("Check is : {}", checkJson);
 					metadata.put(ConsulConstant.MetadataMapKey.CHECK_KEY, checkJson);
@@ -138,8 +128,8 @@ public final class RegistrationUtil {
 		}
 	}
 
-	public static NewService.Check createCheck(Integer port, TsfHeartbeatProperties ttlConfig,
-			TsfDiscoveryProperties properties, TsfCoreProperties tsfCoreProperties) {
+	public static NewService.Check createCheck(Integer port, ConsulHeartbeatProperties ttlConfig,
+			ConsulDiscoveryProperties properties, TsfCoreProperties tsfCoreProperties) {
 		NewService.Check check = new NewService.Check();
 		if (ttlConfig.isEnabled()) {
 			check.setTtl(ttlConfig.getTtl());
@@ -154,12 +144,12 @@ public final class RegistrationUtil {
 		}
 		else {
 			check.setHttp(String.format("%s://%s:%s%s", tsfCoreProperties.getScheme(),
-					AddressUtils.getIpCompatible(properties.getHostname()), port,
+					IPAddressUtils.getIpCompatible(properties.getHostname()), port,
 					properties.getHealthCheckPath()));
 		}
 		check.setInterval(properties.getHealthCheckInterval());
 		check.setTimeout(properties.getHealthCheckTimeout());
-		if (StringUtils.hasText(properties.getHealthCheckCriticalTimeout())) {
+		if (StringUtils.isNotBlank(properties.getHealthCheckCriticalTimeout())) {
 			check.setDeregisterCriticalServiceAfter(properties.getHealthCheckCriticalTimeout());
 		}
 		check.setTlsSkipVerify(properties.getHealthCheckTlsSkipVerify());
@@ -169,7 +159,8 @@ public final class RegistrationUtil {
 	/**
 	 * @return if the management service should be registered with the {@link ServiceRegistry}
 	 */
-	public static boolean shouldRegisterManagement(AutoServiceRegistrationProperties autoServiceRegistrationProperties, TsfDiscoveryProperties properties, ApplicationContext context) {
+	public static boolean shouldRegisterManagement(AutoServiceRegistrationProperties autoServiceRegistrationProperties,
+			ConsulDiscoveryProperties properties, ApplicationContext context) {
 		return autoServiceRegistrationProperties.isRegisterManagement()
 				&& getManagementPort(properties, context) != null
 				&& ManagementServerPortUtils.isDifferent(context);
@@ -178,7 +169,7 @@ public final class RegistrationUtil {
 	/**
 	 * @return the port of the Management Service
 	 */
-	public static Integer getManagementPort(TsfDiscoveryProperties properties, ApplicationContext context) {
+	public static Integer getManagementPort(ConsulDiscoveryProperties properties, ApplicationContext context) {
 		// If an alternate external port is specified, use it instead
 		if (properties.getManagementPort() != null) {
 			return properties.getManagementPort();